analysis-worker/requirements.txt
pika==1.3.2
openai==1.35.7
python-dotenv==1.0.1
httpx==0.27.0
httpcore==1.0.5
PyGithub==2.3.0